No longer serving the firmware hex files
For some as-yet-unknown reason, when the Speck downloads the hex files from the Apache (v2.4) on specksensor.com, it fails. From specksensor.org (on tm1, Apache 2.2), it works.  But Apache 2.4 works when running on Paul's machine.  Whatever.  We're now serving just the hex files from tm1 under the domain hex.specksensor.com, so there's no need for them here.